Removed the set_cancelled call on manager message error handler.
In the main manager receive from workers, if an exception was
generated, the error handler logged some messages and then set the
cancelled flag on the status object.  The status object is never used
again, and this call does not have the effect of cancelling anything
(as is done by a call to MPI_Cancel, but that only works for nonblocking
communications).  I am removing the line.