Update JobController to change interface for kill handling (just set wait_time)
I think the standard way to do things *ought* to be to send a SIGTERM,
wait politely for termination, then send a SIGKILL if the process does
not wrap up before some timeout.  We can do this just by setting a timeout
argument (and use 0 to indicate "SIGKILL at once" and None to indicate
"never SIGKILL").